Separated announce logic into a separate class. Fixed bugs. Probably broke other things.
