+++ src/Neptunium/Data/BuiltinArtists.xml
@@ -99,4 +99,8 @@
   <Artist Name="SISTAR" JPopAsiaUrl="http://www.jpopasia.com/sistar/" FanArtTVUrl="https://fanart.tv/artist/f5bff650-07d0-4f8c-bf48-45497f465b79/sistar/">
     <AltName>씨스타</AltName>
   </Artist>
+  <Artist Name="Aya Hirano" JPopAsiaUrl="http://www.jpopasia.com/ayahirano/" FanArtTVUrl="https://fanart.tv/artist/dfe0486a-d8f7-449a-a1ce-192803b3b537/hirano-aya/">
+    <AltName>Hirano Aya</AltName>
+    <AltName>平野綾</AltName>
+  </Artist>
 </Artists>
